fix docker
