+++ src/models/lstm_model_20240320_13_27.keras
