+++ src/models/lstm_model_20240523_21_11.keras
