+++ src/models/lstm_model_20240524_00_23.keras
