+++ src/models/lstm_model_20240524_16_18.keras
