+++ src/models/lstm_model_20240525_15_25.keras
