+++ src/models/lstm_scaler_20240320_13_27.bin
