+++ src/models/lstm_scaler_20240523_21_11.bin
