+++ src/models/lstm_scaler_20240524_10_49.bin
