+++ src/models/lstm_scaler_20240524_14_16.bin
