+++ src/models/lstm_scaler_20240524_16_18.bin
