+++ src/models/lstm_scaler_20240525_12_44.bin
