+++ src/models/lstm_scaler_20240525_15_25.bin
