--- src/trading_bot_api/Dockerfile
@@ -1,18 +1,0 @@
-# Use an official Python runtime as a parent image
-FROM python:3.11
-
-# Set the working directory in the container
-WORKDIR /usr/src/app
-
-# Copy the current directory contents into the container at /usr/src/app
-COPY . .
-
-# Install any needed packages specified in requirements.txt
-RUN pip install --upgrade pip
-RUN pip install --no-cache-dir -r requirements.txt
-
-# Make port 80 available to the world outside this container
-EXPOSE 80
-
-# Run run_trading_bot.py when the container launches
-CMD ["python", "./run_trading_bot.py"]
