+++ src/trading_bot_api/api_settings.py
@@ -3,7 +3,7 @@ api_key_secret_demo = "wQaYNV1hHLPsdRMl8qB5Ks1e4zeRFXPkceasUgISSLRaehs4SThr7BtNc
 db_conn = "postgresql://db_user:pgpassword123@localhost/opa_db"
 kpi_table = "f_dvkpi"
 kpi_stream_table = "kpi_stream_table"
-klines_stream_table ="f_klines_stream"
+klines_stream_table = "f_klines_stream"
 coin_to_trade = "ETH"
 fiat_curr = "EUR"
 aggregate_trades_stream_table = "f_aggr_trades_stream"
