+++ src/trading_bot_api/requirements.txt
@@ -5,14 +5,9 @@ annotated-types==0.7.0
 anyio==4.4.0
 astunparse==1.6.3
 attrs==23.2.0
-blis==0.7.11
-catalogue==2.0.10
 certifi==2024.6.2
 charset-normalizer==3.3.2
 click==8.1.7
-cloudpathlib==0.18.1
-confection==0.1.5
-cymem==2.0.8
 dateparser==1.2.0
 dnspython==2.6.1
 email_validator==2.1.1
@@ -32,29 +27,20 @@ idna==3.7
 Jinja2==3.1.4
 joblib==1.4.2
 keras==3.3.3
-keras-models==0.0.7
-langcodes==3.4.0
-language_data==1.2.0
 libclang==18.1.1
-marisa-trie==1.2.0
 Markdown==3.6
 markdown-it-py==3.0.0
 MarkupSafe==2.1.5
 mdurl==0.1.2
 ml-dtypes==0.3.2
 multidict==6.0.5
-murmurhash==1.0.10
 namex==0.0.8
 numpy==1.26.4
-opencv-python==4.10.0.82
 opt-einsum==3.3.0
 optree==0.11.0
-orjson==3.10.4
+orjson==3.10.5
 packaging==24.1
 pandas==2.2.2
-pathlib==1.0.1
-pillow==10.3.0
-preshed==3.0.9
 protobuf==4.25.3
 pycryptodome==3.20.0
 pydantic==2.7.4
@@ -69,28 +55,19 @@ PyYAML==6.0.1
 regex==2024.5.15
 requests==2.32.3
 rich==13.7.1
-scikeras==0.13.0
 scikit-learn==1.5.0
 scipy==1.13.1
-sequential==1.0.0
 setuptools==70.0.0
 shellingham==1.5.4
 six==1.16.0
-smart-open==7.0.4
 sniffio==1.3.1
-spacy==3.7.5
-spacy-legacy==3.0.12
-spacy-loggers==1.0.5
 SQLAlchemy==2.0.30
-srsly==2.4.8
 starlette==0.37.2
 tensorboard==2.16.2
 tensorboard-data-server==0.7.2
 tensorflow==2.16.1
 termcolor==2.4.0
-thinc==8.2.4
 threadpoolctl==3.5.0
-tqdm==4.66.4
 typer==0.12.3
 typing_extensions==4.12.2
 tzdata==2024.1
@@ -99,9 +76,7 @@ ujson==5.10.0
 urllib3==2.2.1
 uvicorn==0.30.1
 uvloop==0.19.0
-wasabi==1.1.3
 watchfiles==0.22.0
-weasel==0.4.1
 websockets==12.0
 Werkzeug==3.0.3
 wheel==0.43.0
